pyproject.toml
[project]
name = "fm_eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.0",
    "huggingface-hub>=0.34.1",
    "ipykernel>=6.30.0",
    "ipywidgets>=8.1.7",
    "lightning>=2.5.2",
    "matplotlib>=3.10.3",
    "monai[itk,nibabel]==1.3.2",
    "numpy==1.26.4",
    "open-clip-torch==2.23.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pingouin>=0.5.5",
    "protobuf>=6.31.1",
    "rich>=14.1.0",
    "scikit-learn>=1.7.1",
    "scikit-posthocs>=0.11.4",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.0",
    "tableone>=0.9.5",
    "timm==1.0.19",
    "torch>=2.7.1",
    "transformers>=4.54.0",
    "wandb>=0.21.0",
]
